node:http: skip readStop for upgrade-with-body; move pause tests to their own file
For an accepted Upgrade with a body, the upgrade listener owns the
socket's flow state for the tunnel bytes that follow the body, so
readStop(socket) from the request's push callback (mirroring
parserOnBody) must not touch it. _read() already special-cases upgrade
for the same reason. Without this the tunnel bytes stalled behind a
paused socket on Windows, where do_pause() does not stop kernel reads
and the body fills the Readable buffer in-process.

Move the two new 'pause' tests to their own file so the gate's
with-fix run is not tripped by the pre-existing 'request via http
proxy' failure in node-http.test.ts.
